Simplify router for LensProxy
