+++ packages/core/src/main/router/create-handler-for-route.injectable.ts
@@ -5,41 +5,12 @@
 import { getInjectable } from "@ogre-tools/injectable";
 import type { ServerResponse } from "http";
 import loggerInjectable from "../../common/logger.injectable";
-import { object } from "@k8slens/utilities";
 import type { LensApiRequest, Route } from "./route";
 import { contentTypes } from "./router-content-types";
+import { writeServerResponseFor } from "./write-server-response";
 
 export type RouteHandler = (request: LensApiRequest<string>, response: ServerResponse) => Promise<void>;
 export type CreateHandlerForRoute = (route: Route<unknown, string>) => RouteHandler;
-
-interface LensServerResponse {
-  statusCode: number;
-  content: any;
-  headers: {
-    [name: string]: string;
-  };
-}
-
-const writeServerResponseFor = (serverResponse: ServerResponse) => ({
-  statusCode,
-  content,
-  headers,
-}: LensServerResponse) => {
-  serverResponse.statusCode = statusCode;
-
-  for (const [name, value] of object.entries(headers)) {
-    serverResponse.setHeader(name, value);
-  }
-
-  if (content instanceof Buffer) {
-    serverResponse.write(content);
-    serverResponse.end();
-  } else if (content) {
-    serverResponse.end(content);
-  } else {
-    serverResponse.end();
-  }
-};
 
 const createHandlerForRouteInjectable = getInjectable({
   id: "create-handler-for-route",
@@ -53,25 +24,21 @@ const createHandlerForRouteInjectable = getInjectable({
         const result = await route.handler(request);
 
         if (!result) {
-          const mappedResult = contentTypes.txt.resultMapper({
+          writeServerResponse(contentTypes.txt.resultMapper({
             statusCode: 204,
             response: undefined,
-          });
-
-          writeServerResponse(mappedResult);
+          }));
         } else if (!result.proxy) {
           const contentType = result.contentType || contentTypes.json;
 
           writeServerResponse(contentType.resultMapper(result));
         }
       } catch(error) {
-        const mappedResult = contentTypes.txt.resultMapper({
+        logger.error(`[ROUTER]: route ${route.path}, called with ${request.path}, threw an error`, error);
+        writeServerResponse(contentTypes.txt.resultMapper({
           statusCode: 500,
           error: error ? String(error) : "unknown error",
-        });
-
-        logger.error(`[ROUTER]: route ${route.path}, called with ${request.path}, threw an error`, error);
-        writeServerResponse(mappedResult);
+        }));
       }
     };
   },
